Module physics::misc

Structs

ColData
Data needed to create a 'Collider'.

PhysicObject
Contains all the data needed to create an object in the physic world.

RbData
Data needed to create a 'RigidBody'.

Enums

ShapeType
Different types of shape a PhysicObject can take.

Functions

make_trimesh
Creates the ShapeType::TriMesh associated to the object and return it.

process_shape
Creates and returns a shape needed to create a Collider.